C07/exercise-clinical-feature-store/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
duckdb>=0.9.0
pandera>=0.17.0
pydantic>=2.0.0
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.1.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
